chore: Working on intial setup and what to include
The code changes primarily involve the expansion and refactoring of the Gestalt framework. This includes the addition of new projects, classes, interfaces, and methods, as well as updates to existing files. The changes are designed to enhance the functionality of the Gestalt framework, particularly in relation to ASP.NET MVC, Razor Pages, and SignalR.